Update type of RC measurement timestamps to double, to capture nsec

+++ omf_common/lib/omf_common/dsl/xmpp.rb
@@ -23,13 +23,13 @@ module OmfCommon
         jid = "#{username}@#{server}"
         client.setup(jid, password)
         client.run
-        MPConnection.inject(Time.now.to_i, jid, 'connect') if OmfCommon::Measure.enabled?
+        MPConnection.inject(Time.now.to_f, jid, 'connect') if OmfCommon::Measure.enabled?
       end
 
       # Shut down XMPP connection
       def disconnect
         shutdown
-        OmfCommon::DSL::Xmpp::MPConnection.inject(Time.now.to_i, jid, 'disconnect') if OmfCommon::Measure.enabled?
+        OmfCommon::DSL::Xmpp::MPConnection.inject(Time.now.to_f, jid, 'disconnect') if OmfCommon::Measure.enabled?
       end
 
       # Create a new pubsub topic with additional configuration
@@ -51,7 +51,7 @@ module OmfCommon
       # @param [String] topic Pubsub topic name
       def subscribe(topic, &block)
         pubsub.subscribe(topic, nil, default_host, &callback_logging(__method__, topic, &block))
-        MPSubscription.inject(Time.now.to_i, jid, 'join', topic) if OmfCommon::Measure.enabled?
+        MPSubscription.inject(Time.now.to_f, jid, 'join', topic) if OmfCommon::Measure.enabled?
       end
 
       # Un-subscribe all existing subscriptions from all pubsub topics.
@@ -59,7 +59,7 @@ module OmfCommon
         pubsub.subscriptions(default_host) do |m|
           m[:subscribed] && m[:subscribed].each do |s|
             pubsub.unsubscribe(s[:node], nil, s[:subid], default_host, &callback_logging(__method__, s[:node], s[:subid]))
-            MPSubscription.inject(Time.now.to_i, jid, 'leave', s[:node]) if OmfCommon::Measure.enabled?
+            MPSubscription.inject(Time.now.to_f, jid, 'leave', s[:node]) if OmfCommon::Measure.enabled?
           end
         end
       end
@@ -75,7 +75,7 @@ module OmfCommon
       def publish(topic, message, &block)
         raise StandardError, "Invalid message" unless message.valid?
         pubsub.publish(topic, message, default_host, &callback_logging(__method__, topic, message.operation, &block))
-        MPPublished.inject(Time.now.to_i, jid, topic, message.to_s.gsub("\n",'')) if OmfCommon::Measure.enabled?
+        MPPublished.inject(Time.now.to_f, jid, topic, message.to_s.gsub("\n",'')) if OmfCommon::Measure.enabled?
       end
 
       # Generate OMF related message
@@ -134,7 +134,7 @@ module OmfCommon
       def topic_event(&block)
         guard_block = proc do |event|
           passed = (event.items?) && (!event.delayed?) && event.items.first.payload
-          MPReceived.inject(Time.now.to_i, jid, event.node, event.items.first.payload.to_s.gsub("\n",'')) if OmfCommon::Measure.enabled? && passed
+          MPReceived.inject(Time.now.to_f, jid, event.node, event.items.first.payload.to_s.gsub("\n",'')) if OmfCommon::Measure.enabled? && passed
           passed
         end
         pubsub_event(guard_block, &callback_logging(__method__, &block))
